escritorio/NGROK/requirements.txt
Flask
Flask-Bcrypt
PyJWT
python-dotenv